feat: do not mock vendor settings fetch
