+++ src/types/Env.ts
@@ -24,5 +24,4 @@ export type Env = {
   oauth2ClientFactory: IOAuth2ClientFactory;
   data: DataAdapters;
   sendEmail: (client: Client, email: EmailOptions) => Promise<void>;
-  fetchVendorSettings: (vendor_id?: string) => Promise<VendorSettings>;
 };
